Support some regional emoji flags (closes #157)
This adds support for the regional emoji tag sequences.
Currently the only supported ones are England, Scotland, and Wales.
If more catch on, we can support them easily.

This commit also changes our handling of 'FX' re: #25.
Before, we just followed the formula to return 🄵🅇 from the
enclosed alphanumeric supplement unicode block.
Now we skip it, because we know that won't work.  This allows
country-coder to roll up to 'FR' which does show the French flag.