Add local server
Bun has a builtin http server, so we can use it for local testing.
Just `bun start` or `bun run start` to test country-coder on localhost.

This also uses Bun's HTMLRewriter to rewrite the script src so that
any links to the CDN `dist/*` files get rewritten as local links when
testing this way.