Switch from HTMLRewriter to string replace for rewriting CDN URLs
HTMLRewriter is ok for picking out the `src` of a script tag,
but I found for other projects I want to rewrite links in other
places, like img tags or strings within the actual scripts.

This switches to a more generic string replacement approach that will
find CDN URLs in those other places too.

String replacement should be ok to do here because it will only
be used on html files that easily fit in memory, and none of
these projects are very complicated.